Create a MiniMap (#43)
* create MiniMap

* hud integration

* small ui fixes

* small action fix

* small action fix

* small action fix gradle

* small action fix gradle

* add zoom

* new Textures for Collectables

* Add comprehensive Javadoc comments and fix minor issues

Added Javadoc comments to methods and classes for better code readability and understanding. Also, fixed the typo in the variable name 'aspactRatio' to 'aspectRatio' and made minor changes in other parts of the code, such as renaming variables for better clarity and understanding. Several package-info files were also updated or created to provide high-level information about the packages' contents.