level designer (#46)
* Started Editor

* create grid

* idk

* play with zoom

* create drawings

* straight lines

* lines

* refactored Draw logic to tool

* Shortcuts

* create Square tool

* fileopen

* new file chooser

* settings

* persistent settings

* More settings and persistent Map after exit and Map test mode

* More settings and persistent Map after exit and Map test mode

* Improve code quality and add JavaDoc comments.

* Add import statement in EditorTool class

* Bucket tool

* small fix

* redo undo

* QOL chnages

* javadoc

* Fix map validation to check for Entry instead of Exit

The previous implementation of the MapGenerator was incorrectly checking the grid for an Exit instead of an Entry. This has been corrected to ensure that the map validation checks for the presence of an Entry. This will prevent errors being thrown in the case of a map not having the required Entry point.

* Add error handling to map creation process

The map creation process now includes error handling mechanisms to improve the user experience. When creating or testing a map, error messages will be shown to the user, and the application will not terminate unexpectedly. Also, the "InvalidMapFile" exception import was removed as it is no longer necessary.

* Add error handling to map creation process

The map creation process now includes error handling mechanisms to improve the user experience. When creating or testing a map, error messages will be shown to the user, and the application will not terminate unexpectedly. Also, the "InvalidMapFile" exception import was removed as it is no longer necessary.

* Update editor's shortcuts and initial screen

Added a new shortcut for movement in X direction and improved error handling in shortcuts. Also modified the initial navigation to directly open the editor instead of menu. Scrolling speeds were also adjusted in CanvasInputProcessor.

* last commit