Optimized screen rendering, added window modes and improved save-feature
Several enhancements have been made to the screen rendering process, focusing specifically on the use of texture regions in Wall.java. Adjustments were made to MazeRunnerGame.java to introduce different window modes, involving the set-up of both windowed and fullscreen options. Additionally, modifications were made to enable the saving of the last played map. This enables direct loading of this map for a QuickStart option, improving the user experience.